Rename files and remove useless codes
