+++ Algorithms/src/nattybear/level1/나누어-떨어지는-숫자-배열/Solution.java
@@ -1,0 +1,18 @@
+import java.util.Arrays;
+
+class Solution {
+  public int[] solution(int[] xs, int d) {
+    int[] ys = go(xs, d);
+    if (ys.length == 0)
+      return new int[] {-1};
+    Arrays.sort(ys);
+    return ys;
+  }
+
+  public int[] go(int[] xs, int d) {
+    return Arrays
+      .stream(xs)
+      .filter(x -> x % d == 0)
+      .toArray();
+  }
+}
